refactor : move handleTypeMismatchException()
