+++ src/main/java/kuit/project/beering/controller/advice/FieldValidationControllerAdvice.java
@@ -1,10 +1,12 @@
 package kuit.project.beering.controller.advice;
 
-import kuit.project.beering.util.BaseResponse;
-import kuit.project.beering.util.BaseResponseStatus;
 import kuit.project.beering.dto.common.FieldErrorsDto;
 import kuit.project.beering.dto.common.FieldValidationError;
+import kuit.project.beering.util.BaseResponse;
+import kuit.project.beering.util.BaseResponseStatus;
 import kuit.project.beering.util.exception.validation.FieldValidationException;
+import lombok.extern.slf4j.Slf4j;
+import net.minidev.json.JSONObject;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.FieldError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
@@ -12,9 +14,14 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
+import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
 
+import java.util.HashMap;
 import java.util.List;
 
+import static kuit.project.beering.util.BaseResponseStatus.METHOD_ARGUMENT_TYPE_MISMATCH;
+
+@Slf4j
 @RestControllerAdvice(annotations = RestController.class)
 public class FieldValidationControllerAdvice {
 
@@ -44,5 +51,26 @@ public class FieldValidationControllerAdvice {
                 .build();
 
         return new BaseResponse<>(BaseResponseStatus.INVALID_FIELD, build);
+    }
+
+
+    private static final HashMap<String, String> messageTypeMismatchException = new HashMap<>();
+    static {
+        messageTypeMismatchException.put("date", "date의 format은 yyyy-MM-dd 입니다.");
+    }
+
+    @ExceptionHandler(MethodArgumentTypeMismatchException.class)
+    @ResponseStatus(HttpStatus.METHOD_NOT_ALLOWED)
+    public BaseResponse<?> handleTypeMismatchException(MethodArgumentTypeMismatchException ex) {
+        log.info("MethodArgumentTypeMismatchException");
+
+        String msg = messageTypeMismatchException.get(ex.getName());
+        if(msg == null)
+            msg = ex.getMessage();
+
+        JSONObject result = new JSONObject();
+        result.put(ex.getName(), msg);
+
+        return new BaseResponse<>(METHOD_ARGUMENT_TYPE_MISMATCH, result);
     }
 }
